% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc.R
\docType{data}
\name{pbmc}
\alias{pbmc}
\title{Seurat object of the scRNAseq portion of a 10X single cell immune profiling procedure}
\format{
Seurat object
}
\source{
<https://www.10xgenomics.com/resources/datasets/pbm-cs-from-bal-bc-mice-5-gene-expression-1-standard-3-0-0>
}
\usage{
pbmc
}
\description{
Seurat object of 31053 features across 8000 samples within 1 RNA assay with 2 dimensional reductions of pca and umap calculated.
The object is a result of a rudimentary analysis pipeline of the raw barcode, features, and matrix files.
}
\examples{
  pbmc
}
\keyword{datasets}
