\name{SCIPData-package}
\alias{SCIPData-package}
\alias{SCIPData}
\docType{package}
\title{
\packageTitle{SCIPData}
}
\description{
\packageDescription{SCIPData}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SCIPData}
\packageIndices{SCIPData}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SCIPData}

Maintainer: \packageMaintainer{SCIPData}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
