% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcr_dataframe.R
\docType{data}
\name{tcr_dataframe}
\alias{tcr_dataframe}
\title{all_contig_annotations}
\format{
data.frame
}
\source{
<https://www.10xgenomics.com/resources/datasets/pbm-cs-from-bal-bc-mice-tcr-enrichment-from-amplified-c-dna-1-standard-3-0-0>
}
\usage{
tcr_dataframe
}
\description{
A `data.frame` object of a read-in csv of the `all_contig_annotations.csv` file from 10X's single cell immune profiling.
}
\examples{
  tcr_dataframe
}
\keyword{datasets}
